use multi-thread distribution
